pyproject.toml
[project]
name = "sap-sp-automation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "gitpython>=3.1.45",
    "hdbcli>=2.26.26",
    "openai>=2.9.0",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "rich>=14.2.0",
]

[tool.pytest.ini_options]
markers = [
    "integration: mark test as integration test (deselect with '-m not integration')"
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
